third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

CONSOLE ERROR: Error while parsing the 'suborigin' Content Security Policy directive: No suborigin name specified.

This is a testharness.js-based test.
FAIL Invalid suborigin names Uncaught Error: assert_unreached: Reached unreachable code
Harness: the test ran to completion.
